+++ carapace/cmd/completers.go
@@ -1,0 +1,344 @@
+package cmd
+
+import (
+	basename "github.com/rsteube/carapace-bin/completers/basename_completer/cmd"
+	bat "github.com/rsteube/carapace-bin/completers/bat_completer/cmd"
+	cat "github.com/rsteube/carapace-bin/completers/cat_completer/cmd"
+	chgrp "github.com/rsteube/carapace-bin/completers/chgrp_completer/cmd"
+	chmod "github.com/rsteube/carapace-bin/completers/chmod_completer/cmd"
+	chown "github.com/rsteube/carapace-bin/completers/chown_completer/cmd"
+	chroot "github.com/rsteube/carapace-bin/completers/chroot_completer/cmd"
+	cksum "github.com/rsteube/carapace-bin/completers/cksum_completer/cmd"
+	comm "github.com/rsteube/carapace-bin/completers/comm_completer/cmd"
+	cp "github.com/rsteube/carapace-bin/completers/cp_completer/cmd"
+	csplit "github.com/rsteube/carapace-bin/completers/csplit_completer/cmd"
+	curl "github.com/rsteube/carapace-bin/completers/curl_completer/cmd"
+	cut "github.com/rsteube/carapace-bin/completers/cut_completer/cmd"
+	date "github.com/rsteube/carapace-bin/completers/date_completer/cmd"
+	df "github.com/rsteube/carapace-bin/completers/df_completer/cmd"
+	dir "github.com/rsteube/carapace-bin/completers/dir_completer/cmd"
+	dircolors "github.com/rsteube/carapace-bin/completers/dircolors_completer/cmd"
+	dirname "github.com/rsteube/carapace-bin/completers/dirname_completer/cmd"
+	docker_compose "github.com/rsteube/carapace-bin/completers/docker-compose_completer/cmd"
+	du "github.com/rsteube/carapace-bin/completers/du_completer/cmd"
+	env "github.com/rsteube/carapace-bin/completers/env_completer/cmd"
+	exa "github.com/rsteube/carapace-bin/completers/exa_completer/cmd"
+	expand "github.com/rsteube/carapace-bin/completers/expand_completer/cmd"
+	expr "github.com/rsteube/carapace-bin/completers/expr_completer/cmd"
+	factor "github.com/rsteube/carapace-bin/completers/factor_completer/cmd"
+	fmt "github.com/rsteube/carapace-bin/completers/fmt_completer/cmd"
+	fold "github.com/rsteube/carapace-bin/completers/fold_completer/cmd"
+	git "github.com/rsteube/carapace-bin/completers/git_completer/cmd"
+	gradle "github.com/rsteube/carapace-bin/completers/gradle_completer/cmd"
+	head "github.com/rsteube/carapace-bin/completers/head_completer/cmd"
+	hostid "github.com/rsteube/carapace-bin/completers/hostid_completer/cmd"
+	id "github.com/rsteube/carapace-bin/completers/id_completer/cmd"
+	install "github.com/rsteube/carapace-bin/completers/install_completer/cmd"
+	join "github.com/rsteube/carapace-bin/completers/join_completer/cmd"
+	link "github.com/rsteube/carapace-bin/completers/link_completer/cmd"
+	ln "github.com/rsteube/carapace-bin/completers/ln_completer/cmd"
+	logname "github.com/rsteube/carapace-bin/completers/logname_completer/cmd"
+	ls "github.com/rsteube/carapace-bin/completers/ls_completer/cmd"
+	md5sum "github.com/rsteube/carapace-bin/completers/md5sum_completer/cmd"
+	mkdir "github.com/rsteube/carapace-bin/completers/mkdir_completer/cmd"
+	mkfifo "github.com/rsteube/carapace-bin/completers/mkfifo_completer/cmd"
+	mknod "github.com/rsteube/carapace-bin/completers/mknod_completer/cmd"
+	mv "github.com/rsteube/carapace-bin/completers/mv_completer/cmd"
+	nice "github.com/rsteube/carapace-bin/completers/nice_completer/cmd"
+	nl "github.com/rsteube/carapace-bin/completers/nl_completer/cmd"
+	nohup "github.com/rsteube/carapace-bin/completers/nohup_completer/cmd"
+	od "github.com/rsteube/carapace-bin/completers/od_completer/cmd"
+	paste "github.com/rsteube/carapace-bin/completers/paste_completer/cmd"
+	pathchk "github.com/rsteube/carapace-bin/completers/pathchk_completer/cmd"
+	pinky "github.com/rsteube/carapace-bin/completers/pinky_completer/cmd"
+	pkill "github.com/rsteube/carapace-bin/completers/pkill_completer/cmd"
+	pr "github.com/rsteube/carapace-bin/completers/pr_completer/cmd"
+	printenv "github.com/rsteube/carapace-bin/completers/printenv_completer/cmd"
+	ptx "github.com/rsteube/carapace-bin/completers/ptx_completer/cmd"
+	readlink "github.com/rsteube/carapace-bin/completers/readlink_completer/cmd"
+	rm "github.com/rsteube/carapace-bin/completers/rm_completer/cmd"
+	rmdir "github.com/rsteube/carapace-bin/completers/rmdir_completer/cmd"
+	seq "github.com/rsteube/carapace-bin/completers/seq_completer/cmd"
+	sha1sum "github.com/rsteube/carapace-bin/completers/sha1sum_completer/cmd"
+	shred "github.com/rsteube/carapace-bin/completers/shred_completer/cmd"
+	sleep "github.com/rsteube/carapace-bin/completers/sleep_completer/cmd"
+	sort "github.com/rsteube/carapace-bin/completers/sort_completer/cmd"
+	split "github.com/rsteube/carapace-bin/completers/split_completer/cmd"
+	stat "github.com/rsteube/carapace-bin/completers/stat_completer/cmd"
+	sum "github.com/rsteube/carapace-bin/completers/sum_completer/cmd"
+	sync "github.com/rsteube/carapace-bin/completers/sync_completer/cmd"
+	tac "github.com/rsteube/carapace-bin/completers/tac_completer/cmd"
+	tail "github.com/rsteube/carapace-bin/completers/tail_completer/cmd"
+	tee "github.com/rsteube/carapace-bin/completers/tee_completer/cmd"
+	touch "github.com/rsteube/carapace-bin/completers/touch_completer/cmd"
+	tr "github.com/rsteube/carapace-bin/completers/tr_completer/cmd"
+	tsort "github.com/rsteube/carapace-bin/completers/tsort_completer/cmd"
+	tty "github.com/rsteube/carapace-bin/completers/tty_completer/cmd"
+	uname "github.com/rsteube/carapace-bin/completers/uname_completer/cmd"
+	unexpand "github.com/rsteube/carapace-bin/completers/unexpand_completer/cmd"
+	uniq "github.com/rsteube/carapace-bin/completers/uniq_completer/cmd"
+	unlink "github.com/rsteube/carapace-bin/completers/unlink_completer/cmd"
+	users "github.com/rsteube/carapace-bin/completers/users_completer/cmd"
+	vdir "github.com/rsteube/carapace-bin/completers/vdir_completer/cmd"
+	wc "github.com/rsteube/carapace-bin/completers/wc_completer/cmd"
+	who "github.com/rsteube/carapace-bin/completers/who_completer/cmd"
+	whoami "github.com/rsteube/carapace-bin/completers/whoami_completer/cmd"
+	yes "github.com/rsteube/carapace-bin/completers/yes_completer/cmd"
+)
+
+var completers = []string{
+	"basename",
+	"bat",
+	"cat",
+	"chgrp",
+	"chmod",
+	"chown",
+	"chroot",
+	"cksum",
+	"comm",
+	"cp",
+	"csplit",
+	"curl",
+	"cut",
+	"date",
+	"df",
+	"dircolors",
+	"dir",
+	"dirname",
+	"docker-compose",
+	"du",
+	"env",
+	"exa",
+	"expand",
+	"expr",
+	"factor",
+	"fmt",
+	"fold",
+	"git",
+	"gradle",
+	"head",
+	"hostid",
+	"id",
+	"install",
+	"join",
+	"link",
+	"ln",
+	"logname",
+	"ls",
+	"md5sum",
+	"mkdir",
+	"mkfifo",
+	"mknod",
+	"mv",
+	"nice",
+	"nl",
+	"nohup",
+	"od",
+	"paste",
+	"pathchk",
+	"pinky",
+	"pkill",
+	"pr",
+	"printenv",
+	"ptx",
+	"readlink",
+	"rm",
+	"rmdir",
+	"seq",
+	"sha1sum",
+	"shred",
+	"sleep",
+	"sort",
+	"split",
+	"stat",
+	"sum",
+	"sync",
+	"tac",
+	"tail",
+	"tee",
+	"touch",
+	"tr",
+	"tsort",
+	"tty",
+	"uname",
+	"unexpand",
+	"uniq",
+	"unlink",
+	"users",
+	"vdir",
+	"wc",
+	"whoami",
+	"who",
+	"yes",
+}
+
+func executeCompleter(completer string) {
+	switch completer {
+	case "basename":
+		basename.Execute()
+	case "bat":
+		bat.Execute()
+	case "cat":
+		cat.Execute()
+	case "chgrp":
+		chgrp.Execute()
+	case "chmod":
+		chmod.Execute()
+	case "chown":
+		chown.Execute()
+	case "chroot":
+		chroot.Execute()
+	case "cksum":
+		cksum.Execute()
+	case "comm":
+		comm.Execute()
+	case "cp":
+		cp.Execute()
+	case "csplit":
+		csplit.Execute()
+	case "curl":
+		curl.Execute()
+	case "cut":
+		cut.Execute()
+	case "date":
+		date.Execute()
+	case "df":
+		df.Execute()
+	case "dircolors":
+		dircolors.Execute()
+	case "dir":
+		dir.Execute()
+	case "dirname":
+		dirname.Execute()
+	case "docker-compose":
+		docker_compose.Execute()
+	case "du":
+		du.Execute()
+	case "env":
+		env.Execute()
+	case "exa":
+		exa.Execute()
+	case "expand":
+		expand.Execute()
+	case "expr":
+		expr.Execute()
+	case "factor":
+		factor.Execute()
+	case "fmt":
+		fmt.Execute()
+	case "fold":
+		fold.Execute()
+	case "git":
+		git.Execute()
+	case "gradle":
+		gradle.Execute()
+	case "head":
+		head.Execute()
+	case "hostid":
+		hostid.Execute()
+	case "id":
+		id.Execute()
+	case "install":
+		install.Execute()
+	case "join":
+		join.Execute()
+	case "link":
+		link.Execute()
+	case "ln":
+		ln.Execute()
+	case "logname":
+		logname.Execute()
+	case "ls":
+		ls.Execute()
+	case "md5sum":
+		md5sum.Execute()
+	case "mkdir":
+		mkdir.Execute()
+	case "mkfifo":
+		mkfifo.Execute()
+	case "mknod":
+		mknod.Execute()
+	case "mv":
+		mv.Execute()
+	case "nice":
+		nice.Execute()
+	case "nl":
+		nl.Execute()
+	case "nohup":
+		nohup.Execute()
+	case "od":
+		od.Execute()
+	case "paste":
+		paste.Execute()
+	case "pathchk":
+		pathchk.Execute()
+	case "pinky":
+		pinky.Execute()
+	case "pkill":
+		pkill.Execute()
+	case "pr":
+		pr.Execute()
+	case "printenv":
+		printenv.Execute()
+	case "ptx":
+		ptx.Execute()
+	case "readlink":
+		readlink.Execute()
+	case "rm":
+		rm.Execute()
+	case "rmdir":
+		rmdir.Execute()
+	case "seq":
+		seq.Execute()
+	case "sha1sum":
+		sha1sum.Execute()
+	case "shred":
+		shred.Execute()
+	case "sleep":
+		sleep.Execute()
+	case "sort":
+		sort.Execute()
+	case "split":
+		split.Execute()
+	case "stat":
+		stat.Execute()
+	case "sum":
+		sum.Execute()
+	case "sync":
+		sync.Execute()
+	case "tac":
+		tac.Execute()
+	case "tail":
+		tail.Execute()
+	case "tee":
+		tee.Execute()
+	case "touch":
+		touch.Execute()
+	case "tr":
+		tr.Execute()
+	case "tsort":
+		tsort.Execute()
+	case "tty":
+		tty.Execute()
+	case "uname":
+		uname.Execute()
+	case "unexpand":
+		unexpand.Execute()
+	case "uniq":
+		uniq.Execute()
+	case "unlink":
+		unlink.Execute()
+	case "users":
+		users.Execute()
+	case "vdir":
+		vdir.Execute()
+	case "wc":
+		wc.Execute()
+	case "whoami":
+		whoami.Execute()
+	case "who":
+		who.Execute()
+	case "yes":
+		yes.Execute()
+	}
+}
