+++ carapace/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/rsteube/carapace-bin/cmd"
+import "github.com/rsteube/carapace-bin/carapace/cmd"
 
 func main() {
 	cmd.Execute()
