+++ caraparse/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/rsteube/carapace-bin/caraparse/cmd"
+
+func main() {
+	cmd.Execute()
+}
